myus/pyproject.toml
[project]
name = "myus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "coloredlogs>=15.0.1",
    "dacite>=1.9.2",
    "daphne>=4.2.1",
    "django>=6.0",
    "django-cleanup>=9.0.0",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.1.0",
    "django-environ>=0.12.0",
    "django-import-export>=4.3.14",
    "django-js-asset>=3.1.2",
    "django-ninja>=1.5.1",
    "django-quill-editor>=0.1.42",
    "django-ulid>=0.0.4",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "mypy>=1.19.0",
    "mysqlclient>=2.2.7",
    "pillow>=12.0.0",
    "pycryptodome>=3.23.0",
    "pyjwt>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "python-ffmpeg-video-streaming>=0.1.16",
    "stripe>=14.0.1",
]

[dependency-groups]
dev = ["django-debug-toolbar>=5.1.0", "django-stubs>=5.1.3", "mypy>=1.15.0", "ruff>=0.9.0"]

[tool.mypy]
mypy_path = "."                          # ソースルートパスを指定
python_version = "3.14"                  # Pythonのバージョンを指定
strict = true                            # すべての厳密なチェックを有効化
strict_optional = true                   # Optional型の厳密なチェックを有効化
show_error_codes = true                  # エラーコードを表示
check_untyped_defs = true                # 型アノテーションのない関数の中身もチェック
no_implicit_optional = true              # 暗黙的なOptional型を禁止
warn_return_any = true                   # 戻り値の型がAnyの場合に警告
warn_unused_configs = true               # 使用されていない設定がある場合に警告
warn_redundant_casts = true              # 不要な型キャストがある場合に警告
warn_unused_ignores = true               # 不要な型チェック無視の指定がある場合に警告
warn_no_return = true                    # 関数が値を返さない可能性がある場合に警告
warn_unreachable = true                  # 到達不可能なコードがある場合に警告
disallow_any_unimported = true           # インポートされていない型のAnyを禁止
disallow_untyped_defs = true             # 型アノテーションのない関数定義を禁止
disallow_untyped_decorators = true       # 型アノテーションのないデコレータを禁止
disallow_untyped_calls = true            # 型アノテーションのない関数呼び出しを禁止
disallow_incomplete_defs = true          # 一部のパラメータにのみ型アノテーションがある関数を禁止
disable_error_code = ["import-untyped", "var-annotated"]  # import-untyped, var-annotatedエラーを無視
plugins = ["mypy_django_plugin.main"]    # Django stubsプラグイン

# 特定のモジュールの設定
[[tool.mypy.overrides]]
module = ["django.*", "tests.*"]
ignore_missing_imports = true  # 外部ライブラリの型情報が見つからない場合はエラーを無視

[tool.django-stubs]
django_settings_module = "config.settings.base"


[tool.ruff]
line-length = 200

[tool.ruff.lint]
select = ["I"]  # isortのみ有効

[tool.ruff.lint.isort]
combine-as-imports = true
lines-after-imports = 2                                        # import後の空行
no-lines-before = ["third-party", "first-party", "local-folder"]  # セクション間の改行なし
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
known-first-party = ["api", "app", "config"]
